+++ src/test/exceptions_test.dart
@@ -1,3 +1,4 @@
+import 'package:f_orm_m8_sqlite/exceptions/field_parse_exception.dart';
 import 'package:f_orm_m8_sqlite/orm_m8_generator.dart';
 import 'package:source_gen/source_gen.dart';
 import 'package:test/test.dart';
@@ -38,5 +39,26 @@ void main() {
               '''/*\nException: The DbEntity implementations must be annotated with `@DataTable`!'''),
           true);
     });
+
+    test('Test FieldParseException', () async {
+      expect(
+          () =>
+              simulateFieldParseException(), // throwsA(TypeMatcher<FieldParseException>()))
+          throwsA(allOf(
+              TypeMatcher<FieldParseException>(),
+              predicate((e) =>
+                  e.message == 'Just test' &&
+                  e.toString().contains(
+                      'Exception while parsing field "testField" on "TestModel"!')))));
+    });
   });
+}
+
+void simulateFieldParseException() async {
+  try {
+    throw Exception("FieldParseException test");
+  } catch (exception, stack) {
+    throw FieldParseException("testField", "TestModel",
+        inner: exception, trace: stack, message: "Just test");
+  }
 }