+++ src/test/src/model/no_fields_probe.dart
@@ -1,0 +1,6 @@
+import 'package:f_orm_m8/f_orm_m8.dart';
+
+@DataTable("my_account_related_table")
+class AModelWithoutFields implements DbEntity {
+  int id;
+}